RF: datalad.distribution.remove -> datalad.local.remove
Add transition grease too. And a first dedicated test with 80+% coverage
to eventually replace the ancient uninstall tests.
